+++ plugins/web-interface/deps.ts
@@ -1,3 +1,3 @@
 export { Application, Router } from "https://deno.land/x/oak@v12.6.0/mod.ts"
-export { dirname, fromFileUrl } from "https://deno.land/std@0.193.0/path/mod.ts"
+export { dirname, fromFileUrl } from "https://deno.land/std@0.195.0/path/mod.ts"
 export { Bundlee } from "https://deno.land/x/bundlee@0.9.4/mod.ts"