+++ test/deps.ts
@@ -1,2 +1,2 @@
-export { assert, assertEquals, assertRejects, assertThrows } from "https://deno.land/std@0.193.0/testing/asserts.ts"
-export { assertSpyCall, spy } from "https://deno.land/std@0.193.0/testing/mock.ts"
+export { assert, assertEquals, assertNotEquals, assertRejects, assertThrows } from "https://deno.land/std@0.195.0/testing/asserts.ts"
+export { assertSpyCall, spy } from "https://deno.land/std@0.195.0/testing/mock.ts"